python_app/requirements.txt

gpiozero
paho-mqtt
